REST API >> Created New Endpoints /Login, new URL: /MBA/Chatbot. New Mapper, Class Model, Service and Data Access of a Cargo. Added new @bean in SpringConfig for Cargo

+++ REST-API-MBA/src/main/java/upb/edu/restapimba/Services/UsersBusinessServiceInterface.java
@@ -9,7 +9,7 @@ public interface UsersBusinessServiceInterface
 
     public void test();
     
-    public UserModel getByCode(long id);
+    public UserModel getByCode(int id);
 
     public List<UserModel> getUsers();
